refactor: inline params_access into ublkpp_tgt.cpp, flatten lib/internal/
params_access is the sole consumer of ublkpp_int.hpp; bury it in the one
file that uses it and delete the header. The friend grant in ublk_disk.hpp
remains the visible marker of who holds the privilege.

Move common.hpp from lib/internal/ to lib/ - the extra directory level
added no meaningful boundary since every internal src/ component already
included it.
