+++ example/ublkpp_disk.cpp
@@ -17,48 +17,27 @@
 #include <ublkpp/raid.hpp>
 #include <ublkpp/target.hpp>
 
-#ifdef HAVE_HOMEBLOCKS
-#include <ublkpp/drivers/homeblk_disk.hpp>
-#include <homeblks/home_blks.hpp>
-#include <homeblks/volume_mgr.hpp>
-#endif
 #ifdef HAVE_ISCSI
 #include <ublkpp/drivers/iscsi_disk.hpp>
 #endif
 
-SISL_OPTION_GROUP(
-    ublkpp_disk, (uuid, "", "vol_id", "Volume UUID to use (else random)", ::cxxopts::value< std::string >(), ""),
-    (loop, "", "loop", "Attach a single device 1-to-1", ::cxxopts::value< std::string >(), "<path>"),
-    (raid0, "", "raid0", "Devices for RAID0 device", ::cxxopts::value< std::vector< std::string > >(),
-     "<path>[,<path>,...]"),
-    (raid1, "", "raid1", "Devices for RAID1 device", ::cxxopts::value< std::vector< std::string > >(),
-     "<path>[,<path>,...]"),
-    (raid10, "", "raid10", "Devices for RAID10 device", ::cxxopts::value< std::vector< std::string > >(),
-     "<path>[,<path>,...]"),
-#ifdef HAVE_HOMEBLOCKS
-    (capacity, "", "capacity", "HomeBlks disk capacity GiB", ::cxxopts::value< uint32_t >()->default_value("2"),
-     "<GiB>"),
-    (homeblks_dev, "", "homeblks_dev", "path to the device to run HomeBlocks on", cxxopts::value< std::string >(), ""),
-#endif
-    (stripe_size, "", "stripe_size", "RAID-0 Stripe Size", ::cxxopts::value< uint32_t >()->default_value("131072"), ""),
-    (device_id, "", "device_id", "Recover existing device", cxxopts::value< int32_t >()->default_value("-1"),
-     "<ublkid>"))
+SISL_OPTION_GROUP(ublkpp_disk,
+                  (uuid, "", "vol_id", "Volume UUID to use (else random)", ::cxxopts::value< std::string >(), ""),
+                  (loop, "", "loop", "Attach a single device 1-to-1", ::cxxopts::value< std::string >(), "<path>"),
+                  (raid0, "", "raid0", "Devices for RAID0 device", ::cxxopts::value< std::vector< std::string > >(),
+                   "<path>[,<path>,...]"),
+                  (raid1, "", "raid1", "Devices for RAID1 device", ::cxxopts::value< std::vector< std::string > >(),
+                   "<path>[,<path>,...]"),
+                  (raid10, "", "raid10", "Devices for RAID10 device", ::cxxopts::value< std::vector< std::string > >(),
+                   "<path>[,<path>,...]"),
+                  (stripe_size, "", "stripe_size", "RAID-0 Stripe Size",
+                   ::cxxopts::value< uint32_t >()->default_value("131072"), ""),
+                  (device_id, "", "device_id", "Recover existing device",
+                   cxxopts::value< int32_t >()->default_value("-1"), "<ublkid>"))
 
-#ifdef HAVE_HOMEBLOCKS
-#define HOMEBLKS_OPTIONS , homeblocks, iomgr
-#else
-#define HOMEBLKS_OPTIONS
-#endif
-
-#define ENABLED_OPTIONS logging, ublkpp_tgt, raid1, ublkpp_disk HOMEBLKS_OPTIONS
+#define ENABLED_OPTIONS logging, ublkpp_tgt, raid1, ublkpp_disk
 
 SISL_OPTIONS_ENABLE(ENABLED_OPTIONS)
-
-#ifdef HAVE_HOMEBLOCKS
-#define HOMEBLKS_MODS , HOMEBLOCKS_LOG_MODS, HOMESTORE_LOG_MODS, IOMGR_LOG_MODS
-#else
-#define HOMEBLKS_MODS
-#endif
 
 #ifdef HAVE_ISCSI
 #define ISCSI_MODS , libiscsi
@@ -66,7 +45,7 @@ SISL_OPTIONS_ENABLE(ENABLED_OPTIONS)
 #define ISCSI_MODS
 #endif
 
-SISL_LOGGING_INIT(ublksrv, UBLKPP_LOG_MODS HOMEBLKS_MODS ISCSI_MODS)
+SISL_LOGGING_INIT(ublksrv, UBLKPP_LOG_MODS ISCSI_MODS)
 
 ///
 // Clean shutdown
@@ -86,93 +65,10 @@ static Result _run_target(boost::uuids::uuid const& vol_id, std::shared_ptr< ubl
     return k_target->device_path();
 }
 
-#ifdef HAVE_HOMEBLOCKS
-class UblkPPApplication : public homeblocks::HomeBlocksApplication {
-public:
-    using dev_list_t = std::list< homeblocks::device_info_t >;
-    dev_list_t const _devices;
-    uint32_t const _io_threads;
-    homeblocks::peer_id_t const _id;
-    std::shared_ptr< homeblocks::HomeBlocks > _hb;
-
-    explicit UblkPPApplication(dev_list_t const& dev_list, homeblocks::peer_id_t p = boost::uuids::random_generator()(),
-                               uint32_t const io_threads = 1) :
-            _devices(dev_list), _io_threads(io_threads), _id(p) {}
-    ~UblkPPApplication() override = default;
-
-    // implement all the virtual functions in HomeObjectApplication
-    uint64_t app_mem_size() const override { return 20; }
-    bool spdk_mode() const override { return false; }
-    uint32_t threads() const override { return _io_threads; }
-    dev_list_t devices() const override { return _devices; }
-
-    std::optional< homeblocks::peer_id_t >
-    discover_svc_id(std::optional< homeblocks::peer_id_t > const&) const override {
-        return _id;
-    }
-};
-static std::shared_ptr< UblkPPApplication > _app;
-
-static std::shared_ptr< UblkPPApplication > init_homeblocks(std::string const& hb_dev) {
-    auto const path = std::filesystem::path(hb_dev);
-    if (!std::filesystem::exists(path)) {
-        LOGERROR("Homestore device: {} does not exist!", path.native())
-        return nullptr;
-    }
-
-    auto dev_list = std::list< homeblocks::device_info_t >();
-    dev_list.emplace_back(homeblocks::device_info_t{path.native(), homeblocks::DevType::NVME});
-    auto app = std::make_shared< UblkPPApplication >(dev_list);
-    app->_hb = homeblocks::init_homeblocks(decltype(app)::weak_type(app));
-
-    return app;
-}
-
-static auto create_hb_volume(UblkPPApplication& app, boost::uuids::uuid const& vol_uuid) {
-    auto known_vols = std::vector< homeblocks::volume_id_t >();
-    app._hb->volume_manager()->get_volume_ids(known_vols);
-    if (std::ranges::contains(known_vols, vol_uuid)) return std::error_condition();
-
-    homeblocks::VolumeInfo vol_info;
-    vol_info.page_size = 4 * Ki;
-    vol_info.size_bytes = SISL_OPTIONS["capacity"].as< uint32_t >() * Gi;
-    vol_info.id = vol_uuid;
-    vol_info.name = "ublkpp_disk";
-    LOGINFO("Creating volume: {}", vol_info.to_string());
-
-    return app._hb->volume_manager()
-        ->create_volume(std::move(vol_info))
-        .via(&folly::InlineExecutor::instance())
-        .thenValue([](auto&& e) {
-            if (!e.has_value()) { return std::make_error_condition(std::errc::io_error); }
-            return std::error_condition();
-        })
-        .get();
-}
-#endif
-
 // Return a device based on the format of the input
 // Optional: pass in a unique identifier for metrics tracking
 static std::shared_ptr< ublkpp::ublk_disk > get_driver(std::string const& resource,
                                                        std::string const& metrics_id = "") {
-#ifdef HAVE_HOMEBLOCKS
-    if (0 < SISL_OPTIONS["homeblks_dev"].count()) {
-        if (0 < SISL_OPTIONS["capacity"].count()) {
-            try {
-                auto const vol_uuid = boost::uuids::string_generator()(resource);
-                if (!_app) {
-                    if (_app = init_homeblocks(SISL_OPTIONS["homeblks_dev"].as< std::string >()); !_app) return nullptr;
-                }
-                if (auto e = create_hb_volume(*_app, vol_uuid); e) return nullptr;
-                return std::make_unique< ublkpp::HomeBlkDisk >(vol_uuid, SISL_OPTIONS["capacity"].as< uint32_t >() * Gi,
-                                                               _app->_hb->volume_manager(), 512 * Ki);
-            } catch (std::runtime_error const& e) {}
-        } else {
-            LOGERROR("HomeBlocks device requires --capacity argument!")
-        }
-        return nullptr;
-    }
-#endif
     if (auto path = std::filesystem::path(resource); std::filesystem::exists(path)) {
         return ublkpp::make_fs_disk(path, metrics_id);
     }
@@ -307,9 +203,6 @@ int main(int argc, char* argv[]) {
         ublkpp::ublkpp_tgt::remove(std::move(k_target));
     } else
         s_stop_code.set_value(EIO);
-#ifdef HAVE_HOMEBLOCKS
-    if (_app) _app->_hb->shutdown();
-#endif
     iomanager.stop();
     return exit_future.get();
 }
