+++ src/lib/common.hpp
