+++ src/raid/raid1/bitmap.hpp
@@ -7,7 +7,7 @@
 
 #include "ublkpp/lib/ublk_disk.hpp"
 
-#include "lib/internal/common.hpp"
+#include "lib/common.hpp"
 #include "super_bitmap.hpp"
 
 namespace ublkpp::raid1 {
